Review widget: skip specified number of virtual nodes
The Miscellaneous tab of the review widget now has a new setting:
In-between node step. It defaults to one and can be used to specify how
many virtual nodes can be skipped between two nodes. If the distance
between two nodes would have caused more than this number of virtual
nodes in the focused stack, a virtual node is moved to and the next node
isn't marked as review yet. Once the next node is reached, its review is
confirmed. This setting forces users to sample larger distances.