Force XHR requests to accept all content types as response
This makes XHR requests created through the request queue  include the
'Accept' header field. It is set to '*/*' which means all content types
are requested as response. This is the default in Chromium already, but
some versions of Firefox use this default:

"text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8"

This triggers Django Rest Framework to respond with a webpage (the one
you would see when open the request URL directly in the browser), which
in turn breaks JSON parsing and results in a syntax error being thrown
when Firefox was used. This should be fixed now.